Introduce `SignalDispatchingRepository`
Pull the command-routing cluster up from `ProcessManagerRepository`
into the new abstract base sitting under `EventDispatchingRepository`.
The new class implements `CommandDispatcherDelegate` and declares
`setupCommandRouting()` as `protected abstract`, so every subclass
takes an explicit stance on command routing. `ProcessManagerRepository`
overrides it with a no-op because a Process Manager may handle only
events.
