+++ server/src/main/java/io/spine/server/route/Route.java
@@ -1,11 +1,11 @@
 /*
- * Copyright 2022, TeamDev. All rights reserved.
+ * Copyright 2026, TeamDev. All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- * http://www.apache.org/licenses/LICENSE-2.0
+ * https://www.apache.org/licenses/LICENSE-2.0
  *
  * Redistribution and use in source and/or binary forms, with or without
  * modification, must retain the above copyright notice and the following
@@ -57,6 +57,7 @@ import java.lang.annotation.Target;
  *
  * @see io.spine.server.route.setup.RoutingSetup
  * @see io.spine.server.entity.EventDispatchingRepository#setupEventRouting(EventRouting)
+ * @see io.spine.server.entity.SignalDispatchingRepository#setupCommandRouting(CommandRouting)
  * @see io.spine.server.aggregate.AggregateRepository#setupCommandRouting(CommandRouting)
  */
 @SuppressWarnings("JavadocReference") // We reference protected methods of repository classes.
